handles timezone correctly
Also split command base classes out of commands.py
